Refactor to use gulp-builder
