+++ package.json
@@ -37,14 +37,6 @@
   },
   "devDependencies": {
     "gulp": "^3.9.0",
-    "gulp-jasmine": "^2.2.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-tsc": "^1.2.4",
-    "gulp-typedoc": "^2.0.0",
-    "gulp-typescript": "^2.14.1",
-    "gulp-uglify": "^1.5.1",
-    "object-merge": "^2.5.1",
-    "typedoc": "^0.4.5"
+    "@maxxton/gulp-builder": "1.3.16"
   }
 }
