+++ src/decorators/headers.ts
@@ -1,0 +1,12 @@
+import { RestClient } from "../rest-client";
+
+/**
+ * Set custom headers for a REST method
+ * @param {Object} headersDef - custom headers in a key-value pair
+ */
+export function Headers(headersDef: any) {
+  return function(target: RestClient, propertyKey: string, descriptor: any) {
+    descriptor.headers = headersDef;
+    return descriptor;
+  };
+}
