+++ src/index.ts
@@ -1,0 +1,11 @@
+
+export {RestClient} from './rest-client';
+export {HttpClient} from './abstract/http-client';
+
+export {Client} from './decorators/client';
+export {Headers} from './decorators/headers';
+export {Map} from './decorators/map';
+export {OnEmit} from './decorators/on-emit';
+export {Body, Header, Query, Path} from './decorators/parameters';
+export {MediaType, Produces} from './decorators/produces';
+export {Get, Post, Patch, Put, Delete, Head} from './decorators/request-methods';
